Refactor: Move project structure to CitySecrets and implement basic AppDbContext
