+++ CitySecrets/CitySecrets/Data/AppDbContext
@@ -1,0 +1,23 @@
+using Microsoft.EntityFrameworkCore;
+
+public class AppDbContext : DbContext
+{
+    public AppDbContext(DbContextOptions<AppDbContext> options) : base(options)
+    {
+    }
+
+    public DbSet<Place> Place { get; set; }
+    public DbSet<Review> Reviews { get; set; }
+    public DbSet<User> Users { get; set; }
+    public DbSet<Category> Categories { get; set; }
+    public DbSet<Favorite> Favourites{ get; set; }
+
+    protected override void OnModelCreating(ModelBuilder builder)
+    {
+        base.OnModelCreating(builder);
+
+        builder.Entity<Favorite>()
+            .HasIndex(f => new { f.UserId, f.PlaceId })
+            .IsUnique();
+    }
+}
