--- CitySecrets/CitySecrets/Models/Review.cs
@@ -1,6 +1,0 @@
-﻿namespace CitySecrets.Models
-{
-    public class Review
-    {
-    }
-}
