+++ CitySecrets/CitySecrets/Models/Reviews.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace TempBackend.Models
+namespace CitySecrets.Models
 {
     public class Review
     {
