--- CitySecrets/CitySecrets/Models/User.cs
@@ -1,6 +1,0 @@
-﻿namespace CitySecrets.Models
-{
-    public class User
-    {
-    }
-}
