+++ CitySecrets/CitySecrets/Properties/appsettings.json
@@ -1,0 +1,13 @@
+{
+  "ConnectionStrings": {
+    "DefaultConnection": "Server=DESKTOP-C8PB557\\MSSQLSERVER02;Database=CitySecretsDB;Trusted_Connection=True;TrustServerCertificate=True",
+    "AnotherDatabaseConnection": "Server=my_server_name;Database=AnotherDatabase;User ID=my_user;Password=my_password;"
+  },
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  },
+  "AllowedHosts": "*"
+}
